+++ src/checker.rs
@@ -2,6 +2,8 @@ use std::io::Result;
 use std::process::Command;
 use std::str;
 
+use colored::Colorize;
+
 use crate::utils;
 
 pub fn check_dependencies() -> Result<()> {
@@ -17,15 +19,23 @@ fn check_fastp() {
     let out = Command::new("fastp").arg("--version").output();
 
     match out {
-        Ok(out) => log::info!("[OK]\t{}", str::from_utf8(&out.stderr).unwrap().trim()),
-        Err(_) => log::info!("\x1b[0;41m[NOT FOUND]\x1b[0m\tfastp"),
+        Ok(out) => log::info!(
+            "{:18}: {}",
+            str::from_utf8(&out.stderr).unwrap().trim(),
+            "[OK]".green(),
+        ),
+        Err(_) => log::info!("{:18}: {}", "fastp", "[NOT FOUND]".red()),
     }
 }
 
 fn check_spades() {
     let out = Command::new("spades.py").arg("--version").output();
     match out {
-        Ok(out) => log::info!("[OK]\t{}", str::from_utf8(&out.stdout).unwrap().trim()),
-        Err(_) => log::info!("\x1b[0;41m[NOT FOUND]\x1b[0m\tSPAdes"),
+        Ok(out) => log::info!(
+            "{:18} {}",
+            str::from_utf8(&out.stdout).unwrap().trim(),
+            "[OK]".green(),
+        ),
+        Err(_) => log::info!("{:18}: {}", "SPAdes", "[NOT FOUND]".red()),
     }
 }