+++ src/qc/mod.rs
@@ -4,33 +4,35 @@ mod tag;
 
 use std::path::{Path, PathBuf};
 
+use colored::Colorize;
+
 use crate::qc::parser::RawSeq;
 
 pub fn dry_run(input: &Path, is_id: bool, is_rename: bool) {
     let reads: Vec<RawSeq> = parser::parse_input(input, is_id, is_rename);
     println!();
     reads.iter().for_each(|r| {
-        log::info!("\x1b[0;32mID\t\t: {}\x1b[0m", r.id);
-        log::info!("Read 1\t\t: {}", r.read_1.to_string_lossy());
-        log::info!("Read 2\t\t: {}", r.read_2.to_string_lossy());
+        log::info!("{:18}: {}\x1b[0m", "ID".yellow(), r.id.yellow());
+        log::info!("{:18}: {}", "Read 1", r.read_1.to_string_lossy());
+        log::info!("{:18}: {}", "Read 2", r.read_2.to_string_lossy());
 
         match r.adapter_i7.as_ref() {
             Some(i7) => {
-                log::info!("Adapter i5\t: {}", r.adapter_i5.as_ref().unwrap());
-                log::info!("Adapter i7\t: {}", i7);
+                log::info!("{:18}: {}", "Adapter i5", r.adapter_i5.as_ref().unwrap());
+                log::info!("{:18}: {}", "Adapter i7", i7);
             }
             None => {
                 if r.auto_idx {
-                    log::info!("Adapter\t\t: AUTO-DETECT");
+                    log::info!("{:18}: AUTO-DETECT", "Adapter");
                 } else {
-                    log::info!("Adapter\t\t: {}", r.adapter_i5.as_ref().unwrap());
+                    log::info!("{:18}: {}", "Adapter", r.adapter_i5.as_ref().unwrap());
                 }
             }
         };
 
-        log::info!("Target Dir\t: {}", r.dir.to_string_lossy());
+        log::info!("{:18}: {}", "Target Dir", r.dir.to_string_lossy());
         if is_rename {
-            log::info!("Target fname\t: {}", r.outname.as_ref().unwrap());
+            log::info!("{:18}: {}", "Target fname", r.output_name.as_ref().unwrap());
         }
 
         println!();
@@ -42,8 +44,8 @@ pub fn process_input(
     is_id: bool,
     is_rename: bool,
     params: &Option<String>,
-    outdir: &Option<PathBuf>,
+    output_dir: &Option<PathBuf>,
 ) {
     let reads: Vec<RawSeq> = parser::parse_input(input, is_id, is_rename);
-    runner::clean_reads(&reads, params, outdir);
+    runner::clean_reads(&reads, params, output_dir);
 }